+++ test-suites/variable-debt-token.spec.ts
@@ -1,12 +1,15 @@
 import { expect } from 'chai';
 import { utils } from 'ethers';
-import { DRE, evmRevert, evmSnapshot, impersonateAccountsHardhat } from '../helpers/misc-utils';
-import { getVariableDebtToken } from '../helpers/contracts-getters';
+import { impersonateAccountsHardhat } from '../helpers/misc-utils';
+import { getVariableDebtToken } from '@aave/deploy-v3/dist/helpers/contract-getters';
 import { MAX_UINT_AMOUNT, ZERO_ADDRESS } from '../helpers/constants';
 import { ProtocolErrors, RateMode } from '../helpers/types';
 import { makeSuite, TestEnv } from './helpers/make-suite';
 import { topUpNonPayableWithEther } from './helpers/utils/funds';
 import { convertToCurrencyDecimals } from '../helpers/contracts-helpers';
+import { HardhatRuntimeEnvironment } from 'hardhat/types';
+
+declare var hre: HardhatRuntimeEnvironment;
 
 makeSuite('VariableDebtToken', (testEnv: TestEnv) => {
   const {
@@ -34,7 +37,9 @@ makeSuite('VariableDebtToken', (testEnv: TestEnv) => {
     expect(scaledUserBalanceAndSupplyUser0Before[1]).to.be.eq(0);
 
     // Need to create some debt to do this good
-    await dai.connect(users[0].signer).mint(await convertToCurrencyDecimals(dai.address, '1000'));
+    await dai
+      .connect(users[0].signer)
+      ['mint(uint256)'](await convertToCurrencyDecimals(dai.address, '1000'));
     await dai.connect(users[0].signer).approve(pool.address, MAX_UINT_AMOUNT);
     await pool
       .connect(users[0].signer)
@@ -44,7 +49,7 @@ makeSuite('VariableDebtToken', (testEnv: TestEnv) => {
         users[0].address,
         0
       );
-    await weth.connect(users[1].signer).mint(utils.parseEther('10'));
+    await weth.connect(users[1].signer)['mint(uint256)'](utils.parseEther('10'));
     await weth.connect(users[1].signer).approve(pool.address, MAX_UINT_AMOUNT);
     await pool
       .connect(users[1].signer)
@@ -108,7 +113,7 @@ makeSuite('VariableDebtToken', (testEnv: TestEnv) => {
     // Impersonate the Pool
     await topUpNonPayableWithEther(deployer.signer, [pool.address], utils.parseEther('1'));
     await impersonateAccountsHardhat([pool.address]);
-    const poolSigner = await DRE.ethers.getSigner(pool.address);
+    const poolSigner = await hre.ethers.getSigner(pool.address);
 
     const daiVariableDebtTokenAddress = (
       await helpersContract.getReserveTokensAddresses(dai.address)
@@ -129,7 +134,7 @@ makeSuite('VariableDebtToken', (testEnv: TestEnv) => {
     // Impersonate the Pool
     await topUpNonPayableWithEther(deployer.signer, [pool.address], utils.parseEther('1'));
     await impersonateAccountsHardhat([pool.address]);
-    const poolSigner = await DRE.ethers.getSigner(pool.address);
+    const poolSigner = await hre.ethers.getSigner(pool.address);
 
     const daiVariableDebtTokenAddress = (
       await helpersContract.getReserveTokensAddresses(dai.address)